+++ src/test/kotlin/examples/kotlin/mybatis3/joins/ItemMasterDynamicSQLSupport.kt
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2016-2021 the original author or authors.
+ *    Copyright 2016-2022 the original author or authors.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -15,18 +15,17 @@
  */
 package examples.kotlin.mybatis3.joins
 
-import org.mybatis.dynamic.sql.SqlTable
+import org.mybatis.dynamic.sql.AliasableSqlTable
 import org.mybatis.dynamic.sql.util.kotlin.elements.column
 import java.sql.JDBCType
-import java.util.Date
 
 object ItemMasterDynamicSQLSupport {
     val itemMaster = ItemMaster()
     val itemId = itemMaster.itemId
     val description = itemMaster.description
 
-    class ItemMaster : SqlTable("ItemMaster") {
+    class ItemMaster : AliasableSqlTable<ItemMaster>("ItemMaster", ::ItemMaster) {
         val itemId = column<Int>(name = "item_id", jdbcType = JDBCType.INTEGER)
-        val description = column<Date>(name = "description", jdbcType = JDBCType.DATE)
+        val description = column<String>(name = "description", jdbcType = JDBCType.VARCHAR)
     }
 }